refactoring Globs/TargetAddressInfo into common module
