+++ common/com/twitter/intellij/pants/model/Globs.java
@@ -1,7 +1,7 @@
-// Copyright 2015 Pants project contributors (see CONTRIBUTORS.md).
+// Copyright 2016 Pants project contributors (see CONTRIBUTORS.md).
 // Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-package com.twitter.intellij.pants.service.project.model;
+package com.twitter.intellij.pants.model;
 
 import com.intellij.openapi.util.Condition;
 import com.intellij.openapi.util.text.StringUtil;
