+++ src/com/twitter/intellij/pants/service/project/model/TargetInfoDeserializer.java
@@ -32,9 +32,10 @@ public class TargetInfoDeserializer implements JsonDeserializer<TargetInfo> {
         }
       }
     );
-    final TargetAddressInfo addressInfo = context.deserialize(element, TargetAddressInfo.class);
+    final com.twitter.intellij.pants.model.TargetAddressInfo
+      addressInfo = context.deserialize(element, com.twitter.intellij.pants.model.TargetAddressInfo.class);
     return new TargetInfo(
-      new HashSet<TargetAddressInfo>(Collections.singleton(addressInfo)),
+      new HashSet<com.twitter.intellij.pants.model.TargetAddressInfo>(Collections.singleton(addressInfo)),
       new HashSet<String>(targets),
       new HashSet<String>(libraries),
       new HashSet<String>(excludes),
